// src/types/partials/CapabilityStatementRest.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TMarkdown } from '../simpleTypes/Markdown';
import { TCapabilityStatementSecurity } from '../partials/CapabilityStatementSecurity';
import { TCapabilityStatementResource } from '../partials/CapabilityStatementResource';
import { TCapabilityStatementInteraction1 } from '../partials/CapabilityStatementInteraction1';
import { TCapabilityStatementSearchParam } from '../partials/CapabilityStatementSearchParam';
import { TCapabilityStatementOperation } from '../partials/CapabilityStatementOperation';
import { TCanonical } from '../simpleTypes/Canonical';

export type TCapabilityStatementRest = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    mode: String;
    documentation?: TMarkdown;
    security?: TCapabilityStatementSecurity;
    resource?: TCapabilityStatementResource[];
    interaction?: TCapabilityStatementInteraction1[];
    searchParam?: TCapabilityStatementSearchParam[];
    operation?: TCapabilityStatementOperation[];
    compartment?: TCanonical[];
};